+++ newsfeed/DateExtension.swift
@@ -1,0 +1,55 @@
+//
+//  DateExtension.swift
+//  newsfeed
+//
+//  Created by Anna Oksanichenko on 09.04.2021.
+//
+
+import Foundation
+
+extension Date {
+    
+    func dateTwoDaysAgo() -> String {
+        guard let date = Calendar.current.date(byAdding: .day, value: -2, to: Date()) else {
+            return Date().convertDateToString(from: Date())
+            
+        }
+        let stringDate = convertDateToString(from: date)
+        return stringDate
+    }
+    
+    func convertDateToString(from: Date) -> String {
+       // let now = Date()
+        let formatter = DateFormatter()
+        formatter.timeZone = TimeZone.current
+        formatter.dateFormat = "yyyy-MM-dd"
+        let dateString = formatter.string(from: from)
+        return dateString
+    }
+    
+    func timeAgoDisplay() -> String {
+        
+        let calendar = Calendar.current
+        let minuteAgo = calendar.date(byAdding: .minute, value: -1, to: Date())!
+        let hourAgo = calendar.date(byAdding: .hour, value: -1, to: Date())!
+        let dayAgo = calendar.date(byAdding: .day, value: -1, to: Date())!
+        let weekAgo = calendar.date(byAdding: .day, value: -7, to: Date())!
+        
+        if minuteAgo < self {
+            _ = Calendar.current.dateComponents([.second], from: self, to: Date()).second ?? 0
+            return "right now"
+        } else if hourAgo < self {
+            let diff = Calendar.current.dateComponents([.minute], from: self, to: Date()).minute ?? 0
+            return "\(diff) min ago"
+        } else if dayAgo < self {
+            let diff = Calendar.current.dateComponents([.hour], from: self, to: Date()).hour ?? 0
+            return "\(diff) hrs ago"
+        } else if weekAgo < self {
+            let diff = Calendar.current.dateComponents([.day], from: self, to: Date()).day ?? 0
+            return "\(diff) days ago"
+        }
+        let diff = Calendar.current.dateComponents([.weekOfYear], from: self, to: Date()).weekOfYear ?? 0
+        return "\(diff) weeks ago"
+    }
+    
+}
