+++ newsfeed/NewsFeed/NewsViewModel.swift
@@ -23,10 +23,9 @@ class NewsViewModel {
         case available
         
     }
-       
-    let requestText = "Grammy"
-    var everything = GoogleNewsEverythingRequest(topic: "Sun", dateFrom: Date().currentDate(), dateTo: Date().currentDate(), sortCriteria: .popularity)
     
+    
+    var requestText = "Grammy"
     var modelsForNewsCell = [ModelForNewsCell]()
     var titleForNews = String()
     let googleNewsAPI: GoogleNewsAPI
@@ -74,29 +73,34 @@ class NewsViewModel {
             self.isInternetOn = true
         })
     }
-
+    
     
     
     func showNewsByEverythingRequest(with text: String) {
-        everything.topic = text.isEmpty ? self.requestText : text
+        
+        requestText = text
+        let currentDate = Date().convertDateToString(from: Date())
+        let twoDaysAgo = Date().dateTwoDaysAgo()
+        let newsRequest = GoogleNewsEverythingRequest(topic: requestText, dateFrom: currentDate, dateTo: twoDaysAgo , sortCriteria: .popularity)
         
         modelsForNewsCell = []
         self.dataState = .loading
-        googleNewsAPI.fetchEverythingRequest(googleNewsEverythingRequest: everything) { (response) in
+        googleNewsAPI.fetchNewsByRequest(newsRequest) { (response) in
             
             switch response {
             case .success(let result) :
                 var indexOfAppendingArticle: Int = 0
                 for article in result.articles {
                     let modelForNewsCell = ModelForNewsCell(article: article)
+                    // modelsForNewsCell = []
                     self.modelsForNewsCell.append(modelForNewsCell)
                     indexOfAppendingArticle += 1
-                    if indexOfAppendingArticle > self.everything.pageSize - 1 {
+                    if indexOfAppendingArticle > newsRequest.pageSize - 1 {
                         break
                     }
                 }
                 self.lastUpdate = "Last update: \(String(describing: Date().timeAgoDisplay() ))"
-                self.titleForNews = self.everything.topic
+                self.titleForNews = newsRequest.topic
                 self.delegate?.setTitleForNews(newsTitle: self.titleForNews)
                 self.dataState = .available
                 
@@ -105,7 +109,7 @@ class NewsViewModel {
                 print("NewsViewModel -> showNewsByEverythingRequest -> can`t get successful result frrom response. Error \(error.code): \(error.message)")
                 self.dataState = self.modelsForNewsCell.isEmpty ? .empty : .available
             }
-            self.delegate?.updateDataForShowingNews()
+            
         }
     }
     
@@ -135,41 +139,5 @@ class NewsViewModel {
     
 }
 
-extension Date {
-    
-    func currentDate() -> String {
-        let now = Date()
-        let formatter = DateFormatter()
-        formatter.timeZone = TimeZone.current
-        formatter.dateFormat = "yyyy-MM-dd"
-        let dateString = formatter.string(from: now)
-        return dateString
-    }
-    
-    func timeAgoDisplay() -> String {
-        
-        let calendar = Calendar.current
-        let minuteAgo = calendar.date(byAdding: .minute, value: -1, to: Date())!
-        let hourAgo = calendar.date(byAdding: .hour, value: -1, to: Date())!
-        let dayAgo = calendar.date(byAdding: .day, value: -1, to: Date())!
-        let weekAgo = calendar.date(byAdding: .day, value: -7, to: Date())!
-        
-        if minuteAgo < self {
-            _ = Calendar.current.dateComponents([.second], from: self, to: Date()).second ?? 0
-            return "right now"
-        } else if hourAgo < self {
-            let diff = Calendar.current.dateComponents([.minute], from: self, to: Date()).minute ?? 0
-            return "\(diff) min ago"
-        } else if dayAgo < self {
-            let diff = Calendar.current.dateComponents([.hour], from: self, to: Date()).hour ?? 0
-            return "\(diff) hrs ago"
-        } else if weekAgo < self {
-            let diff = Calendar.current.dateComponents([.day], from: self, to: Date()).day ?? 0
-            return "\(diff) days ago"
-        }
-        let diff = Calendar.current.dateComponents([.weekOfYear], from: self, to: Date()).weekOfYear ?? 0
-        return "\(diff) weeks ago"
-    }
-    
-}
+
